# internship-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Internship E-commerce Backend API
  version: 1.0.0
  description: |
    OpenAPI spec for the Express backend.
    Includes Auth, Products, Cart, Inventory, Orders, and Payments endpoints.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.example.com
    description: Production (replace with your domain)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        email:
          type: string
          nullable: true
        provider:
          type: string
          enum: [password, google, github, guest]
        isGuest:
          type: boolean
      required: [id, uid, provider, isGuest]
    AuthLoginRequest:
      type: object
      properties:
        firebaseIdToken:
          type: string
      required: [firebaseIdToken]
    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with Firebase ID token and receive backend JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user from JWT
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                required: [user]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/hasura-token:
    post:
      tags: [Auth]
      summary: Create Hasura JWT for GraphQL access
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Hasura token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                required: [token]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server/config error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...